% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNAlign.R
\name{CNAlign}
\alias{CNAlign}
\title{CNAlign}
\usage{
CNAlign(
  dat,
  license_path,
  min_ploidy = 1.7,
  max_ploidy = 6.5,
  min_purity = 0.05,
  max_purity = 0.95,
  min_aligned_seg_mb = 5,
  max_homdel_mb = 100,
  delta_tcn_to_int = 0.2,
  delta_tcn_to_avg = 0.1,
  delta_tcnavg_to_int = 0.1,
  delta_mcn_to_int = 0.2,
  delta_mcn_to_avg = 0.1,
  delta_mcnavg_to_int = 0.1,
  mcn_weight = 0.5,
  rho = 1,
  timeout = 120,
  min_cna_segments_per_sample = 5,
  obj2_clonalonly = 0
)
}
\description{
Determine purity and ploidy values for multiple tumor samples with shared ancestry. This function uses the GuRoBi solver to determine purity/ploidy values for each sample that will *maximize* the number of segments with the same (allele-specific) integer copy numbers in at least rho% of samples.
}
